refactor: migrate electron to typescript
