+++ package.json
@@ -2,7 +2,7 @@
   "name": "keepkey-client",
   "version": "0.1.20",
   "private": true,
-  "main": "electron/main.js",
+  "main": "electron/dist/main.js",
   "repository": {
     "url": "https://github.com/keepkey/keepkey-desktop"
   },
@@ -172,7 +172,7 @@
       "electron/assets/light/*",
       "electron/assets/dark/*",
       "electron/icon.png",
-      "electron/main.js",
+      "electron/dist/main.js",
       "package.json"
     ],
     "extends": null,
@@ -212,11 +212,13 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/react-hooks": "^7.0.1",
     "@testing-library/user-event": "^13.2.1",
+    "@types/auto-launch": "^5.0.2",
     "@types/bignumber.js": "^5.0.0",
     "@types/crypto-js": "^4.0.2",
     "@types/d3-array": "^3.0.1",
     "@types/d3-time-format": "^4.0.0",
     "@types/dompurify": "^2.3.2",
+    "@types/electron-is-dev": "^1.1.1",
     "@types/jest": "^26.0.24",
     "@types/lodash": "^4.14.178",
     "@types/node": "^16.11.12",
@@ -230,6 +232,7 @@
     "@types/react-router-dom": "^5.3.2",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
+    "@types/swagger-ui-express": "^4.1.3",
     "babel-plugin-react-remove-properties": "^0.3.0",
     "concurrently": "^6.4.0",
     "cypress": "9.4.1",
